% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcot_colnames.R
\name{fcot_colnames}
\alias{fcot_colnames}
\title{Verify the names of the data.frame}
\usage{
fcot_colnames(data, noms_colonnes, table_orig)
}
\arguments{
\item{data}{tableau}

\item{noms_colonnes}{les noms souhaités}

\item{table_orig}{le nom du tableau}
}
\value{
un tableau de 6 colonnes :
\itemize{
\item test : code du test
\item valeur_test : oui ou non oui correspond à la validation du test
\item table_orig : nom du tableau d'origine
\item test_precis : code du test
\item message : message d'erreur
}
}
\description{
Verify the names of the data.frame
}
\examples{
obj <- data.frame(a=1:5)
fcot_colnames(obj,noms_colonnes=c("a"),table_orig="table test")
}
