% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcot_table.R
\name{fcot_format_table}
\alias{fcot_format_table}
\title{Verify to a data.frame}
\usage{
fcot_format_table(data, table_orig)
}
\arguments{
\item{data}{jeu de donnée à tester}

\item{table_orig}{à voir comment le supprimer}
}
\value{
un tableau de 6 colonnes :
\itemize{
\item test : code du test
\item valeur_test : oui ou non oui correspond à la validation du test
\item table_orig : nom du tableau d'origine
\item test_precis : code du test
\item message : message d'erreur
}
}
\description{
Verify to a data.frame
}
