% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcot_type_character.R
\name{fcot_type_character}
\alias{fcot_type_character}
\title{Check string type of columns}
\usage{
fcot_type_character(data, noms_colonnes, table_orig)
}
\arguments{
\item{data}{tableau}

\item{noms_colonnes}{les nom des colonnes dont on doit vérifier le type}

\item{table_orig}{nom du tableau}
}
\value{

}
\description{
Check string type of columns
}
\examples{
obj <- data.frame(a=1:5)
fcot_type_character(obj,noms_colonnes=c("a"),table_orig="table test")

}
