% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcot_type_numeric.R
\name{fcot_type_numeric}
\alias{fcot_type_numeric}
\title{Check numeric type of columns}
\usage{
fcot_type_numeric(data, noms_colonnes, table_orig)
}
\arguments{
\item{data}{tableau}

\item{noms_colonnes}{les nom des colonnes dont on doit verifier le type}

\item{table_orig}{nom du tableau}
}
\value{

}
\description{
Check numeric type of columns
}
