% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcot_value_empty.R
\name{fcot_value_empty}
\alias{fcot_value_empty}
\title{Check empty values}
\usage{
fcot_value_empty(data, noms_colonnes, table_orig)
}
\arguments{
\item{data}{tableau}

\item{noms_colonnes}{les nom des colonnes dont on doit verifier le type}

\item{table_orig}{nom du tableau}
}
\value{

}
\description{
Check empty values
}
