% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcot_value_interval.R
\name{fcot_value_interval}
\alias{fcot_value_interval}
\title{Check interval values}
\usage{
fcot_value_interval(data, noms_colonnes, mini, maxi, table_orig)
}
\arguments{
\item{data}{tableau}

\item{noms_colonnes}{les nom des colonnes dont on doit verifier le type}

\item{table_orig}{nom du tableau}
}
\value{

}
\description{
Check interval values
}
