% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcot_value_positive.R
\name{fcot_value_positive}
\alias{fcot_value_positive}
\title{Check positive values}
\usage{
fcot_value_positive(data, noms_colonnes, table_orig)
}
\arguments{
\item{data}{tableau}

\item{noms_colonnes}{les nom des colonnes dont on doit verifier le type}

\item{table_orig}{nom du tableau}
}
\value{

}
\description{
Check positive values
}
